fix: 修复登陆时数据解析错误

+++ lib_api/src/main/java/com/imooc/lib_api/model/user/LoginBean.java
@@ -275,7 +275,7 @@ public class LoginBean {
 
         private String detailDescription;
         private boolean followed;
-        private int userId;
+        private long userId;
         private boolean defaultAvatar;
         private String avatarUrl;
         private String nickname;
@@ -323,7 +323,7 @@ public class LoginBean {
             this.followed = followed;
         }
 
-        public int getUserId() {
+        public long getUserId() {
             return userId;
         }
 
@@ -629,7 +629,7 @@ public class LoginBean {
 
         private int refreshTime;
         private String url;
-        private int userId;
+        private long userId;
         private String tokenJsonStr;
         private long id;
         private int type;
@@ -653,7 +653,7 @@ public class LoginBean {
             this.url = url;
         }
 
-        public int getUserId() {
+        public long getUserId() {
             return userId;
         }
